+++ app/main.py
@@ -1,15 +1,17 @@
 import cv2 as cv
 import heapq
 import math
+import model
 import numpy as np
 import os
+import pickle
 import pymongo
 import requests
 import time
 from flask import Flask, json, render_template, request, Response, redirect, send_from_directory, url_for
-from image_similarity_measures.quality_metrics import psnr, rmse, sre
-from multiprocessing import Pool, Process, cpu_count
 from PIL import Image
+from scipy.spatial.distance import cosine
+from threading import Thread
 
 application = Flask(__name__)
 application.config.from_prefixed_env('APP')
@@ -26,12 +28,17 @@ def allowed_file(filename: str) -> bool:
     return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
 
 
-def resize_image(img_name: str, size: int) -> None:
+def resize_image(img_name: str) -> None:
     img = cv.imread(os.path.join(application.config['ORIG_IMAGES'], img_name), cv.IMREAD_COLOR)
     height = img.shape[0]
     width = img.shape[1]
-    preferred_height = size
-    preferred_width = size
+    if height > width:
+        preferred_height = height
+        preferred_width = height
+    else:
+        preferred_height = width
+        preferred_width = width
+
     pad_top = 0
     pad_bot = 0
     pad_left = 0
@@ -39,15 +46,15 @@ def resize_image(img_name: str, size: int) -> None:
 
     if height > width:
         preferred_width = round(preferred_height / height * width)
-        pad_left = math.floor((size - preferred_width) / 2)
-        pad_right = math.ceil((size - preferred_width) / 2)
+        pad_left = math.floor((preferred_height - preferred_width) / 2)
+        pad_right = math.ceil((preferred_height - preferred_width) / 2)
 
     if height < width:
         preferred_height = round(preferred_width / width * height)
-        pad_top = math.floor((size - preferred_height) / 2)
-        pad_bot = math.ceil((size - preferred_height) / 2)
+        pad_top = math.floor((preferred_width - preferred_height) / 2)
+        pad_bot = math.ceil((preferred_width - preferred_height) / 2)
 
-    if height != size or width != size:
+    if height != preferred_height or width != preferred_width:
         img_new = cv.resize(img, (preferred_width, preferred_height))
         img_new_padded = cv.copyMakeBorder(
             img_new,
@@ -58,73 +65,51 @@ def resize_image(img_name: str, size: int) -> None:
             cv.BORDER_CONSTANT,
             value=[255, 255, 255]
         )
-        cv.imwrite(os.path.join(application.config['ORIG_IMAGES'], str(size), img_name), img_new_padded)
-
-    if height == size and width == size:
-        cv.imwrite(os.path.join(application.config['ORIG_IMAGES'], str(size), img_name), img)
+        cv.imwrite(os.path.join(application.config['ORIG_IMAGES'], img_name), img_new_padded)
 
 
-def get_resized_image_path(original_img_name: str, site: str) -> str:
-    if site in ['Lazada', 'Shopee', 'Temu']:
-        return os.path.join(application.config['ORIG_IMAGES'], '700', original_img_name)
-
-    return os.path.join(application.config['ORIG_IMAGES'], '350', original_img_name)
-
-
-def compare_images(original_img_path, compare_img_path, compare_img_name: str) -> dict:
-    original_img = cv.imread(original_img_path)
-    compare_img = cv.imread(compare_img_path)
-
-    return {
-        'img_name': compare_img_name,
-        'psnr': psnr(original_img, compare_img),
-        'rmse': rmse(original_img, compare_img),
-        'sre': sre(original_img, compare_img)
-    }
-
-
-def get_similar_by_metric(compare_results: dict, metric_size=10) -> set:
+def get_similar_by_metric(compare_results: dict) -> set:
     result = set()
 
     for metric, metric_values in compare_results.items():
-        if metric == 'rmse':
-            result.update(set(heapq.nsmallest(metric_size, metric_values, key=metric_values.get)))
+        if metric == 'vgg16-fc2':
+            result.update(set(heapq.nlargest(15, metric_values, key=metric_values.get)))
         else:
-            result.update(set(heapq.nlargest(metric_size, metric_values, key=metric_values.get)))
+            result.update(set(heapq.nlargest(10, metric_values, key=metric_values.get)))
 
     return result
 
 
-def calculate_similarity(original_img_path: str, collection: str, site: str) -> dict:
-    total_result = {'psnr': {}, 'rmse': {}, 'sre': {}}
-    with Pool(processes=cpu_count()) as pool:
-        compare_path = '{}/{}/{}/'.format(application.config['COMPARE_IMAGES'], collection, site)
-        args = [
-            (original_img_path,
-             os.path.join(compare_path, compare_img_name),
-             compare_img_name)
-            for compare_img_name in os.listdir(compare_path)
-        ]
-        result = pool.starmap_async(compare_images, args)
+def calculate_similarity(img_resnet, img_vgg16_flatten, img_vgg16_fc2, collection: str, site: str) -> dict:
+    result = {'resnet-avg-pool': {}, 'vgg16-flatten': {}, 'vgg16-fc2': {}}
+    imgs_comp = get_db()['predictions'].find({'site': site, 'type': collection})
 
-        for value in result.get():
-            total_result['psnr'].update({value['img_name']: value['psnr']})
-            total_result['rmse'].update({value['img_name']: value['rmse']})
-            total_result['sre'].update({value['img_name']: value['sre']})
-        pool.close()
-        pool.join()
+    for img_comp in imgs_comp:
+        result['resnet-avg-pool'].update(
+            {img_comp['image']: cosine_similarity(img_resnet, pickle.loads(img_comp['resnet-avg-pool']))}
+        )
+        result['vgg16-flatten'].update(
+            {img_comp['image']: cosine_similarity(img_vgg16_flatten, pickle.loads(img_comp['vgg16-flatten']))}
+        )
+        result['vgg16-fc2'].update(
+            {img_comp['image']: cosine_similarity(img_vgg16_fc2, pickle.loads(img_comp['vgg16-fc2']))}
+        )
 
-    return total_result
+    return result
+
+
+def cosine_similarity(a, b):
+    return 1 - cosine(a, b)
 
 
 def get_exchange_rate_data(currency='PHP') -> dict:
-    if not is_dump_exist('static/compare_results', currency.lower() + '_rate'):
+    if not is_dump_exist(application.config['EXCHANGE_RATE'], currency.lower() + '_rate'):
         update_exchange_rate_data(currency)
 
-    rate_data = get_dump_data('static/compare_results', currency.lower() + '_rate')
+    rate_data = get_dump_data(application.config['EXCHANGE_RATE'], currency.lower() + '_rate')
 
     if rate_data['time_next_update_unix'] <= int(time.time()) and update_exchange_rate_data(currency):
-        rate_data = get_dump_data('static/compare_results', currency.lower() + '_rate')
+        rate_data = get_dump_data(application.config['EXCHANGE_RATE'], currency.lower() + '_rate')
 
     return rate_data
 
@@ -136,7 +121,7 @@ def update_exchange_rate_data(currency='PHP') -> bool:
     rate_data = json.loads(response.content)
 
     if 'result' in rate_data and rate_data['result'] == 'success':
-        save_dump_data('static/compare_results', currency.lower() + '_rate', rate_data)
+        save_dump_data(application.config['EXCHANGE_RATE'], currency.lower() + '_rate', rate_data)
         return True
 
     return False
@@ -158,6 +143,10 @@ def save_dump_data(path: str, dump_name: str, dump_data) -> None:
         json.dump(dump_data, f)
 
 
+def remove_dump_data(path: str, dump_name: str) -> None:
+    os.remove('{}/{}.json'.format(path, dump_name))
+
+
 @application.route('/original_images/<filename>')
 def get_original_image(filename) -> Response:
     return send_from_directory(application.config['ORIG_IMAGES'], filename)
@@ -176,8 +165,7 @@ def upload_image() -> Response:
         filename = str(int(time.time())) + '.png'
         img = Image.open(file)
         img.save(os.path.join(application.config['ORIG_IMAGES'], filename))
-        resize_image(filename, 700)
-        resize_image(filename, 350)
+        resize_image(filename)
 
         return redirect(url_for('similarity', original_image=filename))
 
@@ -192,6 +180,7 @@ def all_products() -> str:
 @application.route('/similarity')
 def similarity() -> [str, Response]:
     image_name = request.args.get('original_image')
+    recalculate = request.args.get('recalc')
 
     if not isinstance(image_name, str) or not allowed_file(image_name):
         return Response(status=400)
@@ -199,8 +188,11 @@ def similarity() -> [str, Response]:
     if not os.path.isfile(os.path.join(application.config['ORIG_IMAGES'], image_name)):
         return Response(status=404)
 
+    if recalculate:
+        remove_dump_data(application.config['COMPARE_DATA'], image_name + '_log')
+
     if not is_dump_exist(application.config['COMPARE_DATA'], image_name + '_log'):
-        Process(target=api_similarity_calculate, args=[image_name]).start()
+        Thread(target=api_similarity_calculate, name="Downloader", args=[image_name]).start()
 
         return render_template(
             'similar_products.html',
@@ -234,11 +226,14 @@ def results() -> str:
 
     for image_name in os.listdir(application.config['ORIG_IMAGES']):
         image_path = os.path.join(application.config['ORIG_IMAGES'], image_name)
-        if os.path.isfile(image_path):
+        status_path = os.path.join(application.config['COMPARE_DATA'], image_name + '_log.json')
+
+        if os.path.isfile(image_path) and os.path.isfile(status_path):
             table_data.append({
                 'image': image_path,
-                'date': time.ctime(os.path.getmtime(image_path)),
+                'date': time.ctime(os.path.getmtime(status_path)),
                 'url': url_for('similarity', original_image=image_name),
+                'calculate_url': url_for('similarity', original_image=image_name, recalc=True),
                 'status': get_dump_data(application.config['COMPARE_DATA'], image_name + '_log')
             })
 
@@ -276,53 +271,70 @@ def api_all_products() -> Response:
 
 @application.route('/api/similarity/<image_name>/calculate')
 def api_similarity_calculate(image_name: str) -> Response:
-    if not isinstance(image_name, str) or not allowed_file(image_name):
-        return Response(status=400)
+    with application.app_context():
+        if not isinstance(image_name, str) or not allowed_file(image_name):
+            return Response(status=400)
 
-    if not os.path.isfile(os.path.join(application.config['ORIG_IMAGES'], image_name)):
-        return Response(status=404)
+        if not os.path.isfile(os.path.join(application.config['ORIG_IMAGES'], image_name)):
+            return Response(status=404)
 
-    try:
-        collection = 'bags'
-        sites = get_db()[collection].distinct('site')
-        php_rate = get_exchange_rate_data()['conversion_rates']
-        top_similar_images = set()
-        table_data = []
-        save_dump_data(application.config['COMPARE_DATA'], image_name + '_log', 'calculating')
+        try:
+            collection = 'bags'
+            sites = get_db()['predictions'].distinct('site')
+            php_rate = get_exchange_rate_data()['conversion_rates']
+            table_data = []
+            save_dump_data(application.config['COMPARE_DATA'], image_name + '_log', 'calculating')
+            img_resnet = model.get_image_prediction(
+                os.path.join(application.config['ORIG_IMAGES'], image_name),
+                model.get_resnet_model()
+            )
+            img_vgg16_flatten = model.get_image_prediction(
+                os.path.join(application.config['ORIG_IMAGES'], image_name),
+                model.get_vgg16_model()
+            )
+            img_vgg16_fc2 = model.get_image_prediction(
+                os.path.join(application.config['ORIG_IMAGES'], image_name),
+                model.get_vgg16_model('fc2')
+            )
 
-        for site in sites:
-            original_img_path = get_resized_image_path(image_name, site)
-            total_similarity_result = calculate_similarity(original_img_path, collection, site)
-            top_similar_images.update(get_similar_by_metric(total_similarity_result, 25))
+            for site in sites:
+                total_similarity_result = calculate_similarity(
+                    img_resnet,
+                    img_vgg16_flatten,
+                    img_vgg16_fc2,
+                    collection,
+                    site
+                )
+                top_similar_images = get_similar_by_metric(total_similarity_result)
 
-            for similar_image_name in top_similar_images:
-                similar_image_path = '{}/{}/{}'.format(collection, site, similar_image_name)
-                product_data = get_db()[collection].find({'images.path': similar_image_path}).sort([('price', 1)]).limit(1)
+                for similar_image_name in top_similar_images:
+                    similar_image_path = '{}/{}/{}'.format(collection, site, similar_image_name)
+                    product_data = get_db()[collection].find({'images.path': similar_image_path}).sort([('price', 1)]).limit(2)
 
-                for row in product_data:
-                    if next(filter(lambda d: d.get('url') == row['url'], table_data), None) is None:
-                        if row['currency'] != 'PHP':
-                            php_price = round(float(row['price']) / php_rate[row['currency']], 2)
-                        else:
-                            php_price = row['price']
+                    for row in product_data:
+                        if next(filter(lambda d: d.get('url') == row['url'], table_data), None) is None:
+                            if row['currency'] != 'PHP':
+                                php_price = round(float(row['price']) / php_rate[row['currency']], 2)
+                            else:
+                                php_price = row['price']
 
-                        table_data.append({
-                            'image': similar_image_path,
-                            'name': row['name'],
-                            'price': row['price'],
-                            'currency': row['currency'],
-                            'php_price': php_price,
-                            'url': row['url'],
-                            'site': row['site']
-                        })
+                            table_data.append({
+                                'image': similar_image_path,
+                                'name': row['name'],
+                                'price': row['price'],
+                                'currency': row['currency'],
+                                'php_price': php_price,
+                                'url': row['url'],
+                                'site': row['site']
+                            })
 
-        save_dump_data(application.config['COMPARE_DATA'], image_name, table_data)
-        save_dump_data(application.config['COMPARE_DATA'], image_name + '_log', 'finished')
+            save_dump_data(application.config['COMPARE_DATA'], image_name, table_data)
+            save_dump_data(application.config['COMPARE_DATA'], image_name + '_log', 'finished')
 
-        return json.jsonify({'data': table_data})
-    except Exception as e:
-        save_dump_data(application.config['COMPARE_DATA'], image_name + '_log', str(e))
-        return Response(status=400, response=str(e))
+            return json.jsonify({'data': table_data})
+        except Exception as e:
+            save_dump_data(application.config['COMPARE_DATA'], image_name + '_log', str(e))
+            return Response(status=400, response=str(e))
 
 
 @application.route('/api/similarity/<image_name>/result')